feat: fully fledged json parser for Wien metro

+++ src/train_lines/mod.rs
@@ -1,7 +1,7 @@
 pub mod train;
 pub mod train_line;
 
-pub type StationId = u32;
+pub type StationId = usize;
 pub type TrainId = u32;
 pub type LineId = u32;
 pub type Time = u64;
